Move key handle wrapper to a separate module
