+++ src/key.rs
@@ -1,0 +1,42 @@
+//! Cryptographic key handle
+
+use crate::helpers::Handle;
+use std::ptr::null_mut;
+use winapi::shared::bcrypt::*;
+
+/// Cryptographic key handle used in (a)symmetric algorithms
+pub struct KeyHandle {
+    pub(crate) handle: BCRYPT_KEY_HANDLE,
+}
+
+impl KeyHandle {
+    pub fn new() -> Self {
+        Self { handle: null_mut() }
+    }
+}
+
+impl Drop for KeyHandle {
+    fn drop(&mut self) {
+        if !self.handle.is_null() {
+            unsafe {
+                BCryptDestroyKey(self.handle);
+            }
+        }
+    }
+}
+
+impl Default for KeyHandle {
+    fn default() -> Self {
+        KeyHandle::new()
+    }
+}
+
+impl Handle for KeyHandle {
+    fn as_ptr(&self) -> BCRYPT_KEY_HANDLE {
+        self.handle
+    }
+
+    fn as_mut_ptr(&mut self) -> *mut BCRYPT_KEY_HANDLE {
+        &mut self.handle
+    }
+}
